Move finders.dart into /test directory
